Add new detector function.
