+++ crates/cli/Cargo.toml
@@ -28,6 +28,7 @@ moon_node_platform = { path = "../node/platform" }
 moon_node_tool = { path = "../node/tool" }
 moon_pipeline_provider = { path = "../core/pipeline-provider" }
 moon_platform = { path = "../core/platform" }
+moon_platform_detector = { path = "../core/platform-detector" }
 moon_project = { path = "../core/project" }
 moon_project_graph = { path = "../core/project-graph" }
 moon_system_platform = { path = "../system/platform" }
